feat:Navigation에 있던 코드를  AStar 클래스로로 옮기고 Navigation 안내 기능 추가
