+++ docent/app/src/main/java/a2018_hackathon/navigation/MainActivity.java
@@ -2,6 +2,9 @@ package a2018_hackathon.navigation;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
+
+import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
