+++ docent/app/src/main/java/a2018_hackathon/navigation/MapInfo.java
@@ -6,12 +6,33 @@ public class MapInfo {
     //벽
     public static final int block = 0;
 
+    //지도 가로 세로 크기
+    public static  final int map_rows = 7;
+    public static  final int map_cols = 7;
+
+    //작품 정보
     public static final int work_1 = 1;
     public static final int work_2 = 2;
     public static final int work_3 = 3;
     public static final int work_4 = 4;
     public static final int work_5 = 5;
 
-    public static int[][] museum = new int[][]{                                                                                                                                                                                      //
+    //지도 방향
+    public static final double up_bearing=0;
+    public static final double down_bearing=0;
+    public static final double left_bearing=0;
+    public static final double right_bearing=0;
+
+    //노드 사이거리
+    public static final int node_space = 3;
+
+    public static int[][] museum = new int[][]{
+            {-1,0,0,0,0,0,0},
+            {-1,0,0,0,0,0,0},
+            {-1,0,0,0,0,0,0},
+            {-1,0,0,0,0,0,0},
+            {-1,0,0,0,0,0,0},
+            {-1,0,0,0,0,0,0},
+            {0,-1,-1,-1,-1,-1,-1}
     };
 }
